pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webcrawlagent"
version = "0.1.0"
description = "Headless browser crawler with Gemini-powered summaries and PDF reports"
authors = [{ name = "AutoDev", email = "dev@example.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.23",
  "playwright>=1.47",
  "httpx>=0.27",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "jinja2>=3.1",
  "beautifulsoup4>=4.12",
  "fpdf2>=2.7",
  "sse-starlette>=1.6",
  "rich>=13.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "ruff>=0.6"
]

[project.scripts]
webcrawlagent-cli = "webcrawlagent.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]
